refactor: move propertyDTOs to new package
