+++ rentplace/src/main/java/kattsyn/dev/rentplace/services/PropertyService.java
@@ -1,8 +1,8 @@
 package kattsyn.dev.rentplace.services;
 
-import kattsyn.dev.rentplace.dtos.ImageDTO;
-import kattsyn.dev.rentplace.dtos.PropertyCreateEditDTO;
-import kattsyn.dev.rentplace.dtos.PropertyDTO;
+import kattsyn.dev.rentplace.dtos.images.ImageDTO;
+import kattsyn.dev.rentplace.dtos.properties.PropertyCreateEditDTO;
+import kattsyn.dev.rentplace.dtos.properties.PropertyDTO;
 import kattsyn.dev.rentplace.dtos.filters.PropertyFilterDTO;
 import kattsyn.dev.rentplace.entities.Property;
 import org.springframework.web.multipart.MultipartFile;
